resume-analyzer-backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic>=2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.31.0
pdfplumber==0.10.3
python-docx==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.1
email-validator==2.1.0.post1
regex==2023.12.25
pandas==2.2.0
jinja2==3.1.3
numpy==1.26.4
beautifulsoup4==4.12.3
requests-toolbelt==1.0.0
httpx>=0.28.1
aiofiles==23.2.1
asyncpg==0.29.0
greenlet==3.0.3
argon2-cffi==23.1.0
gunicorn==21.2.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
spacy==3.7.4
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
faiss-cpu==1.8.0
transformers==4.38.2
torch --index-url https://download.pytorch.org/whl/cpu
datasets==2.18.0
evaluate==0.4.1
spacy-transformers==1.3.4
xgboost==2.0.3
lightgbm==4.3.0
networkx==3.2.1
matplotlib==3.8.2
plotly==5.19.0
langchain==0.1.14
langchain-community==0.0.31
slowapi==0.1.9
